[refactor] the core registry
